Mass Refactoring
- Split main.go into differnt files
- Make Database drivers an interface
- Make the mdns handling pluggable via functions
- Create binary using the code as a library
- This should make things way easier to test
