+++ Makefile
@@ -7,7 +7,7 @@ help:
 		@echo ""
 
 build: fmt
-		GO15VENDOREXPERIMENT=1 go build -o mdns -ldflags "-X main.builddate=`date -u '+%Y-%m-%d_%I:%M:%S%p'` -X main.gitref=`git rev-parse HEAD`" main.go
+		GO15VENDOREXPERIMENT=1 go build -o mdns -ldflags "-X main.builddate=`date -u '+%Y-%m-%d_%I:%M:%S%p'` -X main.gitref=`git rev-parse HEAD`" cmd/mdns.go
 
 run:
 		./mdns -debug
